Simplifier some variables and added proper values to IntVeto thresholds - after studies of data in Catania

+++ src/libraries/EventBuilder/TEventHeader.h
@@ -64,7 +64,7 @@ public:
 	}
 
 private:
-	int m_eventType;
+	uint8_t m_eventType;
 	int m_runNumber;
 	int m_eventNumber;
 	int m_eventTime; //Unix Timestamp
